Moved DeepDiff::diff to Arr::deepDiff. Deprecated DeepDiff.
