+++ src/Helpers/Arr.php
@@ -133,6 +133,44 @@ class Arr
     }
 
     /**
+     * @internal This may be removed at any point.
+     *
+     * @param array $a
+     * @param array $b
+     *
+     * @return array [key, left, right][]
+     */
+    public static function deepDiff(array $a, array $b)
+    {
+        if (empty($a)) {
+            $a = [];
+        }
+        if (empty($b)) {
+            $b = [];
+        }
+        $keys = array_keys($a + $b);
+        $result = [];
+
+        foreach ($keys as $k) {
+            if (empty($a[$k])) {
+                $l = null;
+            } else {
+                $l = $a[$k];
+            }
+            if (empty($b[$k])) {
+                $r = null;
+            } else {
+                $r = $b[$k];
+            }
+            if ($l != $r) {
+                $result[] = [$k, $l, $r];
+            }
+        }
+
+        return $result;
+    }
+
+    /**
      * Make a simple array consisting of key=>value pairs, that can be used
      * in select-boxes in forms.
      *
